+++ R/ms_options.R
@@ -18,6 +18,8 @@
 #' \item{show_subtitle}{\code{TRUE} Should the subtitle be displayed?}
 #' \item{subtitle}{\code{label} parameter from \code{explainer}.}
 #' \item{ms_title}{Title of the dashboard.}
+#' \item{ms_subtitle}{Subtitle of the dashboard (makes space between the title and line).}
+#' \item{ms_margin_*}{Dashboard margins. Change \code{margin_top} for more \code{ms_subtitle} space.}
 #' \item{margin_*}{Plot margins. Change \code{margin_left} for longer/shorter axis labels.}
 #' \item{w}{\code{420} in px. Inner plot width.}
 #' \item{h}{\code{280} in px. Inner plot height.}
@@ -105,6 +107,9 @@ ms_options <- function(...) {
     show_subtitle = FALSE,
     subtitle = NULL,
     ms_title = NULL,
+    ms_subtitle = NULL,
+    ms_margin_top = 50,
+    ms_margin_bottom = 50,
     margin_top = 50,
     margin_right = 20,
     margin_bottom = 70,